ai-service/requirements.txt
# AI Crime Detection Service Dependencies

# Core ML
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Video Processing
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0
python-multipart>=0.0.6

# Perceptual Hashing
imagehash>=4.3.0

# Gemini VLM
google-genai>=1.51.0

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
aiofiles>=23.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
